Supercycling kludge for use in E3SM.
Version 3, updated relative to e3sm-supercycle-hack-v2.

To use,
    export amb_sc=N   (bash)
    setenv amb_sc N   (csh)
where N>1 for passive tracer supercycling. N<=1 or not setting this environment
variable disables supercycling.

Optionally,
    export amb_subc_params=0 # false
to subcycle ecosys_compute. Default is 1 (true). This feature is new in this
version and is on by default.

Optionally,
    export amb_diag_carbon=1 # true
to compute the carbon mass conservation diagnostic. Default is 0 (false). Each
time step, output looks like this:
    0: amb> C  5.910E+10  1.388E+09  6.048E+10 -5.073E-07  2.238E-05
The first four numbers corresponding to the four terms in the following, and the
fifth is the sum in the following:
   total carbon tendency =
        total tracer tendencies (DIC DOC zooC spC diatC diazC phaeoC spCaCO3)
      + calcToSed + pocToSed
      - carbon surface fluxes   (DIC DOC zooC spC diatC diazC phaeoC spCaCO3)
      + diag_Jint_Ctot
diag_Jint_Ctot is 0 to machine precision because BGC_mod produces balanced
tendencies when flux into the bottom sediment is included. Lines 2 and 3 cancel
the effect of surface and sediment fluxes in line 1.

v3:
* Add ecosys param subcycling.
* Add carbon mass conservation diagnostic.
v2:
* Add flux averaging for BGC riverFlux, iceFlux, depositionFlux.